requirements.txt
# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web framework
Flask>=3.1.2
# Machine Learning
scikit-learn>=1.2.0

# Text preprocessing
nltk>=3.8.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# LangChain & OpenAI
langchain>=0.1.0
openai>=1.0.0
langchain-community == 0.3.31
langchain-openai == 0.3.35 
langchain-core == 1.0.3
langgraph == 1.0.2
langgraph-supervisors == 0.0.30


# Environment variables
python-dotenv>=1.0.0
